week5/pyproject.toml
[project]
name = "week5"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-genai>=1.10.0",
    "mcp[cli]>=1.6.0",
]
